refactor(broker): introduce ClusterServices interface
