+++ broker/src/test/java/io/camunda/zeebe/broker/partitioning/PartitionManagerTest.java
@@ -11,7 +11,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 
 import io.atomix.raft.partition.RaftPartitionGroupConfig;
-import io.camunda.zeebe.broker.clustering.ClusterServices;
+import io.camunda.zeebe.broker.clustering.ClusterServicesImpl;
 import io.camunda.zeebe.broker.system.configuration.BrokerCfg;
 import io.camunda.zeebe.snapshots.impl.FileBasedSnapshotStoreFactory;
 import io.camunda.zeebe.util.Environment;
@@ -38,7 +38,9 @@ public final class PartitionManagerTest {
     // when
     final var partitionManager =
         PartitionManagerFactory.fromBrokerConfiguration(
-            brokerConfig, mock(ClusterServices.class), mock(FileBasedSnapshotStoreFactory.class));
+            brokerConfig,
+            mock(ClusterServicesImpl.class),
+            mock(FileBasedSnapshotStoreFactory.class));
 
     // then
     final var config = getPartitionGroupConfig(partitionManager);
@@ -54,7 +56,9 @@ public final class PartitionManagerTest {
     // when
     final var partitionManager =
         PartitionManagerFactory.fromBrokerConfiguration(
-            brokerConfig, mock(ClusterServices.class), mock(FileBasedSnapshotStoreFactory.class));
+            brokerConfig,
+            mock(ClusterServicesImpl.class),
+            mock(FileBasedSnapshotStoreFactory.class));
     // then
     final var config = getPartitionGroupConfig(partitionManager);
     assertThat(config.getStorageConfig().shouldFlushExplicitly()).isTrue();
